+++ src/modules/Common/components/Anchor/index.ts
@@ -1,0 +1,2 @@
+export { default as Anchor } from './Anchor';
+export { Props } from './types';