+++ src/modules/Common/components/Anchor/types.ts
@@ -1,0 +1,6 @@
+export interface Props { 
+    children: string,
+    href: string,
+    target?: '_blank',
+    color?: string,
+};