+++ src/modules/Popup/components/Footer/Footer.tsx
@@ -1,12 +1,16 @@
 import * as React from 'react';
 import Iframe from 'react-iframe';
 
+import { colors } from '../../../../theme/colors';
 import {
     Container,
     Row,
     Column,
     Center,
 } from '../../../Common/components/Layout/index';
+import { Text } from '../../../Common/components/Text/index';
+import { Anchor } from '../../../Common/components/Anchor/index';
+
 
 import { decorationProps } from './index';
 
@@ -28,24 +32,29 @@ export default class Footer extends React.PureComponent {
 
                     <Row>
                         <Center>
-                            <p>
-                                <a
-                                    href="https://goo.gl/forms/uWbvB9tsbOrbaXmz2"
-                                    target="_blank"
-                                >
-                                    {linkMsg}
-                                </a>
-                            </p>
+                            <Iframe
+                                src="https://ghbtns.com/github-btn.html?user=artem-solovev&repo=gloc&type=star&count=true"
+                                {...decorationProps.githubFrame}
+                            />
+                        </Center>
+                    </Row>
+
+                    <Row>
+                        <Center>
+                            <Anchor
+                                href="https://goo.gl/forms/uWbvB9tsbOrbaXmz2"
+                                target="_blank"
+                                color={colors.grey900}
+                            >
+                                {linkMsg}
+                            </Anchor>
                         </Center>
 
                     </Row>
 
                     <Row>
                         <Center>
-                            <Iframe
-                                src="https://ghbtns.com/github-btn.html?user=artem-solovev&repo=gloc&type=star&count=true"
-                                {...decorationProps.githubFrame}
-                            />
+                            <Text>Made by <Anchor href="http://artemsolovev.com/" target="_blank" color={colors.pink400}>Artem Solovev</Anchor></Text>
                         </Center>
                     </Row>
                 </Column>