+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### 8.2.1 @ 2020-02-02 00:00
+- optimizations
+    - bundle size. From 627KB to 132KB
+    - removed jQuery
+    - webpack optimizations
+- new features
+    - LOC on liked repos
+- design fixes
+- bugfix
+    - fixed LOC for pinned page
+    - fixed LOC for organization page
+
 ### 8.2.0 @ 2020-01-22 00:00
 - LOC for main user page (pinned repos)
 - new design
@@ -12,7 +24,7 @@
 - feature/#52: style separating for loc messages
 - documentation update
 - added support for next languages:
-- - Bulgarian
-- - Polish
+    - Bulgarian
+    - Polish
 - versions bumping for: jquery, tar, fstream, js-yaml, lodash, mixin-deep
 - refactoring